Add factory function handlin the gameboard's cell functionality
The gameboard will be made by cells. Then, a factory function to
create them has been defined.

We only need to associate a value with them. The methods to get
and set such values have been added.

By default such values are binary 0 and 1, but you can specify the
allowed values with the allowedValue array argument. For the
specific game, we will use [0,1,2] with 0: no mark, 1: player 1
mark, 2: player 2 mark.

Also, by default, the value is initialized with the first value in
allowedValue (e.g., 0), but you can optionally choose a different
one.