Add method to make a move to the gameboard object
A move is marking a cell with the value associated with the current
user. As of now, the current user is hardcoded and does not change.
That function returns true if the move is actually performed, and
false otherwise.

A private method to determine whether a move is allowed or not is
also added.