Add method to get the lines of non-empty equal cells
Method 'getArrayOfLinesOfEqualCells' of Gameboard objects retrieves
an array of all the possible equal lines (rows, columns, diagonals).
Each line is an array of references to the gameboard cells.

As of now, cells have just a value as a property. It might be useful
to add an id, row, columns as additional properties, to possibly use
such information.

A non-empty array of all the possible equal lines means that the game
is finished and there is a winner. This prevents some configurations
to appers, eg, two equal lines.