Fix createGameboard interface + minor fixes
Now instead of getting an array of allowed values, the parameters
are the possible player values and the value of an empty cell.

Also, the currentPlayer is no more a property of the gameboard
object (it is a property of the gameController object). It is now
passed to the makeMove method of the gameboard object as an argument.