Create the gameController object handling the game functionalities
This is done with the createGameController factory function.

The methods to initialize a round and make a move in a given game
have been coded. This also includes determining if the move results
in a 'end of game' outcome, which might be a player winning or a tie.

A very simple roundWinner object has been created with the
createRoundWinner factory function, to keep the information of a
new winner.

Finally, the methods to play the game on the console have been
implemented. These have been used to test the methods mentioned
above.

NOTE. In the playMove method of the gameController object, the
actual winning player is not computed yet, but it is hardcoded
to the first player as a placeholder. We have information of the
winning cells and the player value in there, but not on the
specific player object (which contains, eg, its name).
Possible fixes, one of which is to be implemented:
1) store the player index instead of the player value
2) lookup the value in the players list
3) store the player information / a pointer on the cell, instead of
its value.