Fix the cell object
Now there is no more a value associated to the player object.
Instead, a player reference can be assigned to it, which has a getPlayerValue() method to give the symbol to display in the cell.

By default, no player (null) is associated: the empty cell displayed value is given by the parameter emptyCellValue.