Fix winner player retreival in Gameboard's .playMove() method
A .getCellPlayer() method has been added to the cell object.

Then, in the .playMove() method of the Gameboard object, you get
the correct winning player, if there is at least one line with
equal elements (non-empty equalLines array). In that case, all
the cells in the 'lines of cells' elements of  equalLines have the
same player associated with them, by construction.
To get the winning player, get the player of the first of such
cells: equalLines[0][0].getCellPlayer().