Setup the gameboard structure and style
The HTML structure for the header and the gameboard has been setup
and style.

Right now the structure of the gameboard is hardcoded in the html
file. In the future, this will be generated by the JS file, given
the size of the grid.

Note: the grid is not responsive, yet. There are some visualization
problems in landscape mode. These will be fixed in the future.