Add the 'x' and 'o' symbols
Instead of having a text content with 'x' and 'o' symbols, we now
use a specific svg icon. The specific icon representing x and o
marks are shown adding the class 'x' or 'o', respectively, to the
cell.

Note that pointer events on the cells are disabled if a class 'x'
or 'o' are added to the cell.