Add functionality to play a move by clicking/touching an empty cell
The row/column of the cell is retrieved and the .playMove method of
the gameController object is called.

Right now, if the move is succesful, the mark is only set on the cell.
In case of round ending (there is a tie or winner), as of now only a
console message is shown.

Also, right now the event listener is active since the page loading.
In the next version, it should be triggered only when the game starts.