Add a 'start new game' section
A 'start new game' div has been created. As of now, it only has
a 'start you game button' to start a new game, although this
functionality is not implemented yet.

In the future, this 'start new game' section will contain the game
settings, like the players names selection.

The content that was previously inside the main div has now been
wrapped inside a 'playing' div.

Then, main has only two child: the 'start new game' and 'playing'
view. This is to make it easy to hide one of the two sections,
based on the playing status (todo).