Add 'end game' button in the 'round outcome' div
When a round ends, the player should be able to continue to the
next round, or end the game and return to the initial view. Then,
a button to do that has been added in the 'round outcome' div.

Note that the button has only been added to the page, but is not
functional yet.