Add functionality to adapt the gameboard size to its container size
We want to adapt the gameboard size based on the gameboard container.
In particular, it has to keep the 1 aspect ratio, and fill the
smallest between the height and width of the gameboard container.

A resize observer has been setup to adapt the gameboard size based
on the size of the gameboard container div.

Indeed, a resize observer can react to changes of some associated div.

See https://web.dev/articles/resize-observer