Fix the cells clicking behaviour
Previously, the cells, when on active state, were scaled down.
This prevented the click event to be captured when the click point
was inside the cell on non-active state, and outside it when on
active.

Now a non-scalable cell div wrapper has been added (whose class is .cell-clickable-area).

Then, the gameboard structure becomes:
    gameboard > cell-clickable-area > cell

gameboard has a click event listener, but because of bubbling,
in the event handler, e.target is the top element present that
capture the click.

If e.target is a cell, play the move on that cell. Otherwise, a
cell might still have been selected: in that case, e.target is
a cell-clickable-area.