Add possibility to have a winning length smaller than gamebaord size
Let size be the gameboard size. So far, a player would win if it
marks 'size' cells in line (row, column or diagonal).

In a large 4x4/5x5/... scenario, the more the size increases, the
less the possiblity of winning are. Indeed, it is easier to block
a win of the opponent and many games end in a tie.

By adding the possibility to win if a player marks winLen cells in
line, with winLen <= size, the game becomes more interesting, as
there are more possiblities to win for both players.

Note that so far winLen has been hard coded. A UI selection is not
added yet.