Add proficient and advanced ai players
These two new levels of ai become proactive.

The fork opportunities are taken into account. These are moves that
allow for a win on the successive move of the same player, after
the opponent makes its move.

Let F (forking score) be the number of ways in which a player could
win after a possible move.
- If F=1 there is a potential win for the ai, that however could be
  blocked by the opponent.
- If F>1, the ai has a guaranteed win, because the opponent can
  block only one of them.

Then, the ai should analyze all the potential moves and compute
their forking score {F1,F2,...}. Then, the suggested move is:
- A random move among the ones with F>1 (to secure a win), if
  present,
- Or a random move among the ones with F==1, if present; altough
  this does not neccessarily secure a win, but just a potential
  win,
- Otherwise, the move should use another heuristic (eg, a random
  move)

A similar discussion can be made for the opponents's fork
opportunities, which must be blocked.

Then, for the proficient ai:
- If the ai player can win now, make the move to win,
- else if the opponent can win on the next move, make the move to
  block it,
- else if the ai can create a fork opportunity, create it
- else, make a random move.

For the advanced ai:
- If the ai player can win now, make the move to win,
- else if the opponent can win on the next move, make the move to
  block it,
- else if the ai can create a fork opportunity, create it,
- else if the opponent can create a fork opportunity, block it,
- else, make a random move.