Add Expert and Master AI players.
Both these new players use the minmax algorithm to determine the
best move.

The implementation is based on:
https://alialaa.com/blog/tic-tac-toe-js-minimax

The Master AI explores the whole solution tree (full depth). In
optimal conditions, this ensures that the ai always wins or ties.

The Expert AI has a lower skill level. Instead of exploring the
whole solution tree, it stops after 3 (for the 3x3 grid) or 4
(for the 4x4, 5x5, 6x6 grids). This makes it not perfect and
beatable.

NOTE
For a 3x3 grid it is viable to explore the whole solution space.
However, for larger gameboards, this is not possible due to
the very large nodes to explore.

Currently the program gets stuck for boards larger than 3x3, so
Master AI should be not be used for the 4x4, 5x5, 6x6 boards.

Several optimizations will be introduced to get a (possibly
suboptimal) move in such cases. For instance:
1) reducing the depth of search (at least for the first few
   iterations);
2) using alpha-beta pruning;
3) possibly exploiting the symmetry of the boards.

A better heuristic minmax score should also be used for better
performance when a maxDepth is used.