Add maximum depth computation for minmax method getBestMove().
A rough maximum number of nodes of the search tree to explore is
set.

For a 3x3 grid, the maximum number of nodes to explore is (3*3)!,
ie, 362880 (altough the actual number of nodes is actually smaller
due to the early terminal conditions that don't occur at full
depth).

Since a full search tree can be explored in a 3x3 grid in less
than a second, the maximum number of nodes to explore is set to
500000.

Then, let N be the number of empty cells in a gameboard. The maximum
depth is computed as the maximum M such that N*(N-1)*...*(N-M) is
smaller than the maximum number of nodes to explore (500000).

If the depth of the tree is explicitly limited (eg, for the Expert
AI), the minimum between the two maximum depth is used.