Add gameboard heuristic score to be used in minmax search
So far, in the minmax search method, when the max depth had been
reached without a terminating condition, the gameboard score was
assigned a generic 0. Now an heuristic has been introduced to
estimate the heuristic score of that gameboard, to improve
performance.

This is particularly useful for the case is which the search has
to be stopped at a given depth for computational reasons, eg, on
 larger boards.

Such heuristic score is based on: https://stackoverflow.com/questions/77446399/how-to-properly-implement-minimax-ai-for-tic-tac-toe

For each line (row, column, diagonal), consider the winning
opportunities, ie, lines in which no opponent mark is set.
For each winning opportunity line, cound the number of marks of
the current player in there.
Then, the heuristic score (for a given player) is the sum of the
square of these values.
The overall gameboard heuristic score is the heuristic score of
the current player minus the one of the opponent.

Note that empty lines are ignored, as they are opportunities for
both players.

Also, note that if winLen<size, there can be more than one
'sub-line' in each line, so each of them must be considered as
a separate one.

Currently, the computation of the ai next move takes more time
with this modification. An optimization of the code has to be
performed.